# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiseAuVert API",
        "description": "API documentation for MiseAuVert",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://app.miseauvert.fr",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/animaux": {
            "get": {
                "tags": [
                    "Animaux"
                ],
                "summary": "Liste des animaux de l'utilisateur connecté",
                "operationId": "97fa5f84fd3fe9674a5e115fabbb838a",
                "responses": {
                    "200": {
                        "description": "Liste des animaux"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Animaux"
                ],
                "summary": "Créer un animal",
                "operationId": "038572e69f0a88b048bb654a883cab68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "espece": {
                                        "type": "string"
                                    },
                                    "race": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "poids": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "carnet_vaccination": {
                                        "type": "boolean"
                                    },
                                    "vaccin_a_jour": {
                                        "type": "boolean"
                                    },
                                    "vermifuge_a_jour": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Animal créé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez le token obtenu via /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}